[flash] make mup 64bit safe (stupid, long is 32bit on our system not 64bit)

+++ tdt/flash/common/mup.src/crc32.h
@@ -10,6 +10,6 @@
 
 #include <stdint.h>
 
-uint64_t crc32 (uint8_t *data, uint32_t len);
+uint32_t crc32 (uint8_t *data, uint32_t len);
 
 #endif /* CRC32_H_ */
